vars: use matching cast on both sides of comparison

+++ vars_linux.go
@@ -90,7 +90,7 @@ func probeEfivarfs() bool {
 	if err := varsStatfs(efivarfsPath(), &st); err != nil {
 		return false
 	}
-	if uint32(st.Type) != unix.EFIVARFS_MAGIC {
+	if uint(st.Type) != uint(unix.EFIVARFS_MAGIC) {
 		return false
 	}
 	return true
